requirements.txt
pandas
pyyaml
pytest
transformers
datasets
accelerate
torch
openai
